add example command with own class
some people prefer to have each command in its own class so I wanted to give them the option to do that.
I don't like that style but others do so I added a way to do that. This time I actually implement something I myself don't like
because this is supposed to be a template for everyone and not just me.